start of refactor to move handler code
