+++ handlers/index.js
@@ -1,0 +1,2 @@
+exports.events = require ('./events.js');
+exports.help = require ('./help.js');
